backend/requirements.txt
fastapi==0.131.0
uvicorn[standard]==0.41.0
python-dotenv==1.2.1
pydantic-settings==2.13.1
supabase==2.28.0
groq==1.0.0
python-multipart==0.0.22
httpx===0.28.1
